Website Overhaul Part 4: Build Script
`build-site.el` is an Emacs Lisp script that converts content files into
a "publish"-able format. Most functions and options used are explained
in the official documentation.[1]

`build.sh` calls the script, and then strips away every occurrences of
`file:` protocol text in the output html files. Without the last
command, the links in anchor or image tags would not become the correct
form. I looked online for ways to handle this on the org-mode side, but
apprently it cannot be done without reimplementing the link behavior in
emacs lisp[1], which I decided not to do for now.

[1]: https://orgmode.org/manual/Publishing.html
[2]: https://linevi.ch/en/org-link-extra-attrs.html